+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.0.1-1.19] - 2022-07-19
+### Added
+- Added a config option to control if anvil exclusive enchantments (e.g. sharpness on axes) are allowed to be applied in the infuser
+- This is enabled by default for the advanced infuser
+### Fixed
+- Fixed the infuser always allowing to apply enchantments that usually require an anvil
+
 ## [v4.0.0-1.19] - 2022-07-19
 - Ported to Minecraft 1.19
 - Split into multi-loader project